+++ gdsfactory/routing/validation.py
@@ -3,6 +3,7 @@ from warnings import warn
 import numpy as np
 
 from gdsfactory.component_reference import ComponentReference
+from gdsfactory.config import CONF
 from gdsfactory.port import Port
 from gdsfactory.snap import snap_to_grid
 from gdsfactory.typings import Route
@@ -15,9 +16,9 @@ def validate_connections(
     Validates that a set of Routes indeed connects the port-pairs listed in ports1 and ports2. If the Routes form valid connections between ports1 and ports2, the original Routes will be returned. If not, a RouteWarning will be raised, and a set of error traces will be returned instead.
 
     Args:
-        ports1: the list of starting ports
-        ports2: the list of ending ports
-        routes: the list of Route objects, purportedly between ports1 and ports2
+        ports1: the list of starting ports.
+        ports2: the list of ending ports.
+        routes: the list of Route objects, purportedly between ports1 and ports2.
 
     Returns:
         A list of Routes. If the input routes are valid, they will be returned as-is. Otherwise, a list of error traces will be returned and a RouteWarning will be raised.
@@ -42,9 +43,9 @@ def make_error_traces(
     Creates a set of error traces showing the intended connectivity between ports1 and ports2. The specified message will be included in the RouteWarning that is raised.
 
     Args:
-        ports1: the list of starting ports
-        ports2: the list of ending ports
-        message: a message to include in the RouteWarning that is raised
+        ports1: the list of starting ports.
+        ports2: the list of ending ports.
+        message: a message to include in the RouteWarning that is raised.
 
     Returns:
         A list of Routes (the error traces).
@@ -56,7 +57,7 @@ def make_error_traces(
     error_routes = []
     for port1, port2 in zip(ports1, ports2):
         path = gf.path.Path([port1.center, port2.center])
-        error_component = gf.path.extrude(path, layer="ERROR_PATH", width=1)
+        error_component = gf.path.extrude(path, layer=CONF.layer_error_path, width=1)
         error_ref = ComponentReference(error_component)
         error_route = Route(
             references=[error_ref], ports=list(error_ref.ports.values()), length=np.nan