+++ gdsfactory/technology/klayout_tech.py
@@ -10,6 +10,7 @@ from pydantic import BaseModel, ConfigDict, Field
 
 from gdsfactory.config import PATH
 from gdsfactory.technology import LayerStack, LayerViews
+from gdsfactory.technology.xml_utils import make_pretty_xml
 from gdsfactory.typings import PathType
 
 try:
@@ -88,7 +89,6 @@ class KLayoutTechnology(BaseModel):
             mebes_config: A dictionary specifying the KLayout mebes reader config.
 
         """
-        from gdsfactory.utils.xml_utils import make_pretty_xml
 
         d25_filename = d25_filename or f"{self.name}.lyd25"